added new component Layer and added the Repo layer to the Main.js Component
